+++ src/main/kotlin/net/ossrs/ideasrs/explorer/SrsExplorerFactory.kt
@@ -3,16 +3,11 @@
 
 package net.ossrs.ideasrs.explorer
 
-import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.ActionManager
-import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.ui.SimpleToolWindowPanel
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
-import com.intellij.ui.components.panels.NonOpaquePanel
-import com.intellij.util.ui.UIUtil
 
 import net.ossrs.ideasrs.SrsBundle
 
@@ -34,21 +29,5 @@ class SrsExplorerFactory : ToolWindowFactory, DumbAware {
 
     override fun init(toolWindow: ToolWindow) {
         toolWindow.stripeTitle = SrsBundle.message("explorer.title")
-    }
-}
-
-class SrsExplorerToolWindow(project: Project): SimpleToolWindowPanel(true, true), Disposable {
-    init {
-        background = UIUtil.getTreeBackground()
-        setContent(NonOpaquePanel())
-    }
-
-    override fun dispose() {
-    }
-
-    companion object {
-        fun getInstance(project: Project): SrsExplorerToolWindow {
-            return ServiceManager.getService(project, SrsExplorerToolWindow::class.java)
-        }
     }
 }